+++ addons/wiki/tests/test_models.py
@@ -1,8 +1,12 @@
 import pytest
+from urllib import quote
 
 from modularodm.exceptions import ValidationValueError
-from osf_tests.factories import NodeFactory
+
 from addons.wiki.models import NodeWikiPage
+from addons.wiki.tests.factories import NodeWikiFactory
+from osf_tests.factories import NodeFactory, UserFactory, ProjectFactory
+from tests.base import OsfTestCase
 
 pytestmark = pytest.mark.django_db
 
@@ -43,3 +47,35 @@ class TestNodeWikiPageModel:
         # Simulate a deleted page by not adding ver to
         # node.wiki_pages_current and node.wiki_pages_versions
         assert ver.is_current is False
+
+
+class TestNodeWikiPage(OsfTestCase):
+
+    def setUp(self):
+        super(TestNodeWikiPage, self).setUp()
+        self.user = UserFactory()
+        self.project = ProjectFactory(creator=self.user)
+        self.wiki = NodeWikiFactory(user=self.user, node=self.project)
+
+    def test_factory(self):
+        wiki = NodeWikiFactory()
+        assert wiki.page_name == 'home'
+        assert wiki.version == 1
+        assert wiki.content == 'Some content'
+        assert bool(wiki.user)
+        assert bool(wiki.node)
+
+    def test_url(self):
+        assert self.wiki.url == '{project_url}wiki/home/'.format(project_url=self.project.url)
+
+    def test_url_for_wiki_page_name_with_spaces(self):
+        wiki = NodeWikiFactory(user=self.user, node=self.project, page_name='Test Wiki')
+        url = '{}wiki/{}/'.format(self.project.url, quote(wiki.page_name))
+        assert wiki.url == url
+
+    def test_url_for_wiki_page_name_with_special_characters(self):
+        wiki = NodeWikiFactory(user=self.user, node=self.project)
+        wiki.page_name = 'Wiki!@#$%^&*()+'
+        wiki.save()
+        url = '{}wiki/{}/'.format(self.project.url, quote(wiki.page_name))
+        assert wiki.url == url
